make seed an int

+++ shaders/effects/classicNoisedeck/background/glsl/background.glsl
@@ -10,7 +10,7 @@ precision highp float;
 precision highp int;
 
 uniform float time;
-uniform float seed;
+uniform int seed;
 uniform vec2 resolution;
 uniform int backgroundType;
 uniform float rotation;
